added token authentication and updated test to reflect this

+++ controllers/blogs.js
@@ -1,4 +1,5 @@
 const blogsRouter = require('express').Router()
+const jwt = require('jsonwebtoken')
 const Blog = require('../models/blog')
 const User = require('../models/user')
 
@@ -9,11 +10,17 @@ blogsRouter.get('/', async (request, response) => {
 
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
+  const token = request.token
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+
+  if(!token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
 
   if(!body.author || !body.title)
-    response.status(400).end()
+    response.status(400).json({error: 'missing author or title'})
 
-  const user = await User.findOne({})
+  const user = await User.findById(decodedToken.id)
 
   const blog = new Blog({
     author: body.author,
@@ -23,14 +30,30 @@ blogsRouter.post('/', async (request, response) => {
     user: user._id
   })
 
-  const result = await blog.save()
-  response.status(201).json(result)
+  const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+  response.status(201).json(savedBlog)
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
   const id = request.params.id
-  await Blog.findByIdAndRemove(id)
-  response.status(204).end()
+  const token = request.token
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+
+  const userId = decodedToken.id
+  const blog = await Blog.findById(id)
+
+  if(!token || !decodedToken.id)
+    return response.status(401).json({error: 'token missing or invalid'})
+
+  if(userId === blog.user.toString()) {
+    await Blog.findByIdAndRemove(id)
+    response.status(204).end()
+  } else {
+    response.status(401).json({error: 'this user cannot delete this blog'})
+  }
+  
 })
 
 blogsRouter.put('/:id', async (request, response) => {
